apps/speech-python/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-multipart>=0.0.12
httpx>=0.27.0
prometheus-client>=0.21.0
structlog>=24.4.0
haystack-ai>=2.10.0
huggingface-hub>=0.26.0
sentence-transformers>=3.3.0
onnxruntime>=1.19.0
